Fixed bug with real world value mappings
RWVMs can either be shared or per-frame, so I added some logic where a
single RWVM object will be assigned to the shared FG, but if multiple
RWVMs are passed it goes to per-frame FGs.

Found an issue with Double Float Pixel Data which results in the
attribute being stored as a multivalue attr.